% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-metrics.R
\name{DIC}
\alias{DIC}
\alias{DIC.Asymmfit}
\title{Compute DIC}
\usage{
DIC(x, ...)

\method{DIC}{Asymmfit}(x, ...)
}
\arguments{
\item{x}{An Asymmfit object obtained with 'fit_model()'}

\item{...}{Ignored}
}
\value{
A list containing DIC
}
\description{
Compute DIC
}
\examples{
\dontrun{
fit <- fit_model(...)
DIC(fit)
}


}
