% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-item-icc.R
\name{item_icc}
\alias{item_icc}
\title{Plot the Item Characteristic Curve}
\usage{
item_icc(x, item = 1, theta_lim = c(-12, 12))
}
\arguments{
\item{x}{An object of class \code{"Asymmfit"} (from \code{\link{fit_model}}).}

\item{item}{An integer specifying the item to plot. Default value to 1}

\item{theta_lim}{A vector that defines the limits of the x-axis}
}
\value{
A ggplot object showing the Item Characteristic Curve
}
\description{
Plot the Item Characteristic Curve
}
\examples{

dat <- simData(n=10,k=10,seed=123,model_type="LPE")

if (instantiate::stan_cmdstan_exists()) {

fit <- fit_model(data = dat$df, mod = "LPE2", iter_sampling = 100, iter_warmup = 100,
chains=2, parallel_chains = 2, seed = 123)

item_icc(fit, item = 2)}


}
