% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-item-info.R
\name{item_info}
\alias{item_info}
\title{Item Information Function}
\usage{
item_info(x, item = 1, theta_lim = c(-14, 14))
}
\arguments{
\item{x}{An object of class \code{"Asymmfit"} (from \code{\link{fit_model}})}

\item{item}{An integer specifying the item to plot. Default value to 1}

\item{theta_lim}{A vector that defines the limits of the x-axis}
}
\value{
A ggplot object showing the Item Information
}
\description{
Plot the item information of a specified item from a simdata or Asymmfit object
}
\examples{

dat <- simData(j=10,k=10,seed=123,model_type="LPE")
if (instantiate::stan_cmdstan_exists()) {

fit <- fit_model(data = dat$df, mod = "LPE2", iter_sampling = 50, iter_warmup = 100,
chains=2, parallel_chains = 2, seed = 123)

item_info(fit, item = 2)}
}
