% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-metrics.R
\name{loo}
\alias{loo}
\alias{loo.Asymmfit}
\title{Compute loo}
\usage{
loo(x, ...)

\method{loo}{Asymmfit}(x, ...)
}
\arguments{
\item{x}{An Asymmfit object obtained with 'fit_model()'}

\item{...}{Arguments of the loo function}
}
\value{
A list containing loo
}
\description{
Compute loo
}
\examples{
\dontrun{
fit <- fit_model(...)
loo(fit)
}


}
