% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-Asymmfit.R
\name{plot.Asymmfit}
\alias{plot.Asymmfit}
\title{Plot HPD interval alongside the median of a parameter of the fitted model}
\usage{
\method{plot}{Asymmfit}(x, variable = "b", ci = 0.95, ...)
}
\arguments{
\item{x}{An Asymmfit object created with the fit_model function}

\item{variable}{A string variable indicating the parameter of the fitted model these should be: "a", "lambda" or "b"}

\item{ci}{Value or vector of probabilities for the credible interval (between 0 and 1)}

\item{...}{Arguments passed to `bayestestR::hdi`}
}
\value{
A ggplot object showing the HPD intervals and posterior median of the parameter
}
\description{
Plot HPD interval alongside the median of a parameter of the fitted model
}
\examples{

\dontrun{
fit <- fit_model(...)
plot(x = fit,variable = "b",ci = 0.95)
}
}
