% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-traceplot.R
\name{plot_trace}
\alias{plot_trace}
\alias{plot_trace.Asymmfit}
\title{Plot Traceplots of posterior draws}
\usage{
plot_trace(x, pars = "b", ...)

\method{plot_trace}{Asymmfit}(x, pars = "b", ...)
}
\arguments{
\item{x}{An object of class \code{"Asymmfit"} obtained from \code{\link{fit_model}}.}

\item{pars}{Character or vector string specifying the parameter name to plot (e.g., \code{"b"},\code{"b[1]"})}

\item{...}{Additional arguments passed to \code{bayesplot::mcmc_trace}}
}
\value{
A traceplot (via \code{bayesplot::mcmc_trace})
}
\description{
Plot Traceplots of posterior draws
}
\examples{

\dontrun{
fit <- fit_model(...)
plot_trace(fit, pars = "b")
}


}
