% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-summary-code.R
\name{summary.Asymmfit}
\alias{summary.Asymmfit}
\title{Summary of fitted model}
\usage{
\method{summary}{Asymmfit}(object, variable = c("b", "theta"), ci = 0.95, ...)
}
\arguments{
\item{object}{An Asymmfit object obtained with the fit_model function}

\item{variable}{A vector string referring to the variables to be added to the summary}

\item{ci}{Value or vector of probabilities for the credible interval (between 0 and 1)}

\item{...}{Arguments passed to the CmdStanR summary method}
}
\description{
Summary of fitted model
}
