% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-metrics.R
\name{WAIC}
\alias{WAIC}
\alias{WAIC.Asymmfit}
\title{Compute WAIC}
\usage{
WAIC(x, ...)

\method{WAIC}{Asymmfit}(x, ...)
}
\arguments{
\item{x}{An Asymmfit object obtained with 'fit_model()'}

\item{...}{Arguments passed to}
}
\value{
A list containing WAIC
}
\description{
Compute WAIC
}
\examples{
\dontrun{
fit <- fit_model(...)
WAIC(fit)
}


}
