% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-summary-code.R
\name{code}
\alias{code}
\alias{code.Asymmfit}
\title{Stan code of the model fitted}
\usage{
code(x, ...)

\method{code}{Asymmfit}(x, ...)
}
\arguments{
\item{x}{An Asymmfit object obtained with 'fit_model()'}

\item{...}{Ignored}
}
\value{
stan model code
}
\description{
Stan code of the model fitted
}
\examples{
\dontrun{
fit <- fit_model(...)
code(fit)
}


}
