% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_model}
\alias{fit_model}
\title{Fit symmetric IRT (1PL, 2PL) models and asymmetric IRT models (LPE, RLPE, 1PLPE, 1PRLPE)}
\usage{
fit_model(data, mod, ...)
}
\arguments{
\item{data}{A binary matrix with entries 0 or 1, representing item responses}

\item{mod}{A character string refering to the model to be used to fit the data}

\item{...}{Additional arguments passed to \code{cmdstanr}'s \code{$sample()} method}
}
\value{
An `Asymmfit` object (a list) containing:
\itemize{
  \item \code{output}: A `CmdStanR` object.
  \item \code{data}: The data used to fit the model.
  \item \code{model_type}: The model used in the argument.
}
}
\description{
This function fits some widely known models such as 1PL and 2PL models and also
some asymmetric models like LPE, RLPE, 1PLPE, 1PRLPE. For models such as the LPE and RLPE
there are three different priors to use for the lambda parameter.
}
\details{
The model is selected by specifying it as a string. For example:
\itemize{
  \item \code{"LPE_log"} fits the LPE model and includes the log-likelihood in generated quantities
  to further compute WAIC, loo and DIC
  \item \code{"LPE"} fits the LPE model without the likelihood in generated quantities
}


Models available to fit are

\itemize{
     \item \code{"LPE1"} or \code{"LPE1_log"}
     \item \code{"LPE2"} or \code{"LPE2_log"}
     \item \code{"LPE3"} or \code{"LPE3_log"}
     \item \code{"RLPE1"} or \code{"RLPE1_log"}
     \item \code{"RLPE2"} or \code{"RLPE2_log"}
     \item \code{"RLPE3"} or \code{"RLPE3_log"}
     \item \code{"2PL"}  or \code{"2PL_log"}
     \item \code{"1PL"}  or \code{"1PL_log"}
     \item \code{"1PLPE"}  or \code{"1PLPE_log"}
     \item \code{"1PRLPE"}  or \code{"1PRLPE_log"}

}


Use \code{methods(class = "Asymmfit")} to see methods applicable to this class.
To access further information about these methods use \code{?method_name.Asymmfit}
(e.g \code{?summary.Asymmfit})
}
\examples{

data <- simData(j=20,k=10,seed=123,model_type = "LPE")
if (instantiate::stan_cmdstan_exists()) {
fit <- fit_model(data = data$df, mod = "LPE2_log", iter_sampling = 50, iter_warmup = 50,
chains=2, parallel_chains = 2, seed = 123)

summary(fit)

code(fit) #Stan code

item_icc(fit,1) #Plot of ICC for item 1

item_info(fit,1) #Plot of Item Information Function for item 1


}


}
