% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-simdata.R
\name{plot.simdata}
\alias{plot.simdata}
\title{Plot of scores and item parameters from simulated data}
\usage{
\method{plot}{simdata}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"simdata"} created by [simdata()]}

\item{...}{Ignored}
}
\value{
no returns
}
\description{
This function generates a total scores plot and a boxplot of the item parameters
(`a`, `b`, and `λ`) from a `simdata` object
}
\examples{
data <- simData(j=20,k=500,seed=123,model_type = "LPE")
plot(data)

}
