% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-simdata.R
\name{print.simdata}
\alias{print.simdata}
\title{Print summary of simulated dataset or a binary matrix}
\usage{
\method{print}{simdata}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"simdata"} or a binary matrix containing only zeros and ones, with rows representing individuals and columns representing items}

\item{...}{Ignored}
}
\value{
A matrix with item-level descriptive statistics:
\itemize{
  \item \code{mean}: Item mean
  \item \code{sd}: Item standard deviation
  \item \code{proportion_0}: Proportion of 0 responses
  \item \code{proportion_1}: Proportion of 1 responses
  \item \code{kappa}: Index to conclude imbalanced data, where values > 0.2 indicate unbalanced items
}
The matrix is returned invisibly. A summary is printed to the console
}
\description{
This function computes some descriptive statistics (e.g mean, sd, proportions of 1's and 0's)
of the generated dataset with the simData function or a binary matrix only containing ones and zeros. Kappa's index is computed by following the works
of Bazán et. al (2025), where an item is considered unbalanced if kappa > 0.2
}
\examples{
\dontrun{
fit <- fit_model(...)
print(fit)
}
}
