% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generation.R
\name{simData}
\alias{simData}
\title{Simulate a dataset from a RLPE, LPE, 1PLPE, or 1PRLPE model}
\usage{
simData(j, k, seed, model_type)
}
\arguments{
\item{j}{Integer. Number of items}

\item{k}{Integer. Number of individuals}

\item{seed}{Integer. Random seed for reproducibility}

\item{model_type}{A string that specifies the model (RLPE, LPE, 1PLPE, or 1PRLPE)}
}
\value{
An object of class \code{"simdata"}: a list with components:
\describe{
  \item{items_param}{Item parameters: a, b, and lambda}
  \item{theta}{Latent traits}
  \item{df}{Binary response matrix}
}
}
\description{
This function simulates data and returns a list of the model parameters and a dataframe,
its rows refer to individuals whereas columns to items. The data is simulated from the LPE (Logistic positive exponential)
family of models (e.g RLPE, LPE, 1PRLE, or 1PLPE)
based on the works of Samejima (1999). For more details, see the works of Bazan et al. (2023)
}
\details{
Use \code{methods(class = "simdata")} to see methods applicable to this class.
To access further information about these methods use \code{?method_name.Asymmfit}
(e.g \code{?summary.Asymmfit})
}
\examples{
data <- simData(j=40,k=1000,seed=123,model_type = "RLPE")
plot(data)

}
